add error log
updating angular 11 to latest is blocked

https://github,com/angular/angular-cli mentions this in https://github.com/angular/angular-cli/issues/9029 which directs us to https://github.com/angular/angular-cli/issues/9020 but all of these are for old versions of angular

here is the error text

- Generating browser application bundles (phase: setup)...
Unknown error from PostCSS plugin. Your current PostCSS version is 8.3.11, but postcss-preset-env uses 7.0.39. Perhaps this is the source of the error below.
Unknown error from PostCSS plugin. Your current PostCSS version is 8.3.11, but postcss-preset-env uses 7.0.39. Perhaps this is the source of the error below.
✔ Browser application bundle generation complete.
✔ Browser application bundle generation complete.